+++ src/tcpFiles.h
@@ -16,6 +16,6 @@
 char data[DATA_SIZE];
 
 void send_file(int fd, char* fname, int sp);
-void recv_file(int fd, char* fname, int fsize, char* initial_data, int initial_data_size, int dup);
+void recv_file(int fd, char* fname, long fsize, char* initial_data, int initial_data_size, int dup);
 
 #endif
